defuse test 12
There is disagreement between system whether the aggregation
in the query

select (select sum(x)) from (values(1),(2)) s(x)

should be evaluated in the inner query or in the outer query.
My reading of the SQL standard says that an aggregation function
is evaluated for every input row, but systems like PostgreSQL
interpret that differently and instead conceptially rewrite the query as

select (select sum_x)
from (
    select sum(x) as sum_x
    from (values(1),(2)) s(x)
) t

For now test 12 is formulated in a way that returns the same result
for both interpretation, this requires more thoughts.